Add files via upload
Toggle Menu: The toggleMenu function toggles the visibility of the navigation menu by adding/removing the "show" class to the navbar's list of links.
Carousel Function: The moveSlide function moves through slides of a carousel on the shop page by adjusting the currentSlide index. The updateCarousel function updates the carousel's position using CSS transform. Automatic slide transitions are set to happen every 5 seconds with setInterval.

Shop Page Dynamic Content: The generateShop function dynamically generates and displays products in the shop, based on an array shopItemsData. Each product can be added to the cart by clicking the "Add to Cart" button.

Cart Management:
Add to Cart: The addToCart function adds a selected product to the cart (stored in localStorage), either incrementing the quantity if the item already exists or adding it with a quantity of 1 if it's new.
Update Cart Icon: The calculation function updates the cart icon with the total number of items currently in the cart.
Notifications: A notification appears when an item is added to the cart, which fades out after 3 seconds.
The code makes use of localStorage to persist the shopping cart data, ensuring that the cart contents are retained even when the page is reloaded.